requirements.txt
absl-py==0.11.0
altgraph==0.17
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==0.15.8
asgiref==3.2.10
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
black==20.8b1
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
chardet==3.0.4
click==7.1.2
cmake==3.18.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
curses @ file:///C:/Users/ABDUL%20BASID/Downloads/curses-2.2.1%2Butf8-cp38-cp38-win_amd64.whl
cycler==0.10.0
Cython==0.29.21
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
dlib==19.21.0
dominate==2.5.1
entrypoints==0.3
face-alignment==1.0.1
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
hyperlink==19.0.0
idna==2.10
imageio==2.9.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
make==0.1.6.post1
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.4
notebook==6.1.4
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.52
Pygments==2.6.1
PyHamcrest==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
smmap==3.0.4
sortedcontainers==2.2.2
sqlparse==0.4.1
stickytape==0.1.15
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.8.13
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.1.0
